Fix kazoo-haproxy script
This script was causing an error on Debian 8:

    kazoo-haproxy[50720]: /usr/sbin/kazoo-haproxy: 51: [: root: unexpected operator

The reason for the error is that the script is run under `/bin/sh`,
and the line with the error was

    if [ "$(whoami)" == "${USER}" ]; then

In `sh`, the `[` operator is a synonym for `test`. The above statement
is therefore equivalent to

    if test "$(whoami)" == "${USER}"; then

The problem is that the `sh` version of `test` - at least, in Debian 8,
in which `/bin/sh` is a softlink to `/bin/dash` - requires  a single `=`
for string comparisons; the `==` operator is only supported by `bash`
`test` (although see below).

It is interesting to note that when tested in macOS 10.11.6 and CentOS
7.5, this was not treated as an error. In CentOS 7, `/bin/sh` is
softlinked to `/bin/bash`, so that kind of makes sense, but it's not in
macOS. In any case, it appears that Debian `dash` sticks more closely to
the original Bourne shell version of `test` than the other two OSes.

Probably the best fix is to use the `=` operator instead, as follows:

    if [ "$(whoami)" = "${USER}" ]; then

which is equivalent to

    if test "$(whoami)" = "${USER}"; then

and is most likely valid syntax on all `sh` versions.

This commit changes all the uses of the `==` operator to `=`.