Fix some portability issues
There are two portability issues addressed in this commit:

1. Specifying the shebang as `#!/bin/sh` and then using Bashisms
2. A clash between the meaning of `nc` on Debian and CentOS

Running `systemctl start kazoo-haproxy` on Debian 8 resulted in this
error:

    /usr/sbin/kazoo-haproxy: 51: [: root: unexpected operator

The reason for the error is that the script's shebang is `#!/bin/sh`,
which has different behavior on Debian 8 compared to CentOS 7.

In CentOS 7, `/bin/sh` is symlinked to `/bin/bash`, but on Debian 8,
it's symlinked to `/bin/dash`, which behaves more like the original
Bourne shell, `sh`.

`sh` (and `dash`) supports `=`, but not `==`, as a string comparison
operator, hence the error.

The choices considered for this fix were:

1. Change all occurrences of `==` to`=`
2. Change the shebang to `#!/bin/bash`

This commit changes the shebang.

An earlier version of this PR chose option #1 above, but that was before
discovering that the rest of the script had many Bashisms.

- On CentOS, `nc` is another way of running `ncat`.
- On Debian, `nc` is `netcat`, a totally different utility.

Fortunately, both respond to `ncat`, so this commit changes all uses of
`nc` to `ncat`.