Fix some portability issues
Summary
=======

There are three portability issues addressed in this commit:

1. Specifying the shebang as `#!/bin/sh` and then using Bashisms
2. Mixing Bashisms and sh
2. A clash between the meaning of `nc` on Debian and CentOS

Bashisms
--------
Running `systemctl start kazoo-haproxy` on Debian 8 resulted in this
error:

    /usr/sbin/kazoo-haproxy: 51: [: root: unexpected operator

The reason for the error is that the script's shebang is `#!/bin/sh`,
which has different behavior on Debian 8 compared to CentOS 7.

In CentOS 7, `/bin/sh` is symlinked to `/bin/bash`, but on Debian 8,
it's symlinked to `/bin/dash`, which behaves more like the original
Bourne shell, `sh`.

See [HTMBSWID][How to make Bash scripts work in dash] for detail.

`sh` (and `dash`) supports `=`, but not `==`, as a string comparison
operator, hence one error.

Furthermore, `bash` does not allow `==`

This part of the commit does the following:

1. Change all occurrences of `==` to`=`
2. Change the shebang to `#!/bin/bash`

`nc` vs `ncat`
--------------

- On CentOS, `nc` is another way of running `ncat`.
- On Debian, `nc` is `netcat`, a totally different utility.

Debian `ncat` seems to hang when used in `kazoo-haproxy`. After some
debugging and research, replacing the use of

    ncat -U ${SOCKET}

with

    socat stdio unix-connect:${SOCKET}

seems to work perfectly in both Debian and CentOS, so this commit adopts
the use of `socat`.

[HTMBSWID]: http://mywiki.wooledge.org/Bashism